requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
pyzbar
azure-ai-formrecognizer
dotenv